OrienGo-AI/requirements.txt
Flask==3.1.2
pandas==2.1.3
python-jobspy==1.1.82
requests==2.32.4